requirements.txt
openai>=1.54.3
requests>=2.32.0
python-dotenv>=1.0.1
